+++ src/lib/services/searchService.ts
@@ -20,7 +20,6 @@ import {
 } from "lib/utils";
 
 import { useBlockData } from "./blockService";
-import { useCodeDataByCodeId } from "./codeService";
 import { queryContract } from "./contract";
 import { useModuleByAddressLcd } from "./move/moduleService";
 import { useAddressByICNSName, useICNSNamesByAddress } from "./nameService";
@@ -29,6 +28,7 @@ import { usePoolByPoolId } from "./poolService";
 import { useProposalData } from "./proposalService";
 import { useTxData } from "./txService";
 import { useValidatorData } from "./validatorService";
+import { useCodeDataByCodeId } from "./wasm/code";
 
 export type SearchResultType =
   | "Code ID"